+++ examples/kitchen-sink/template-engine.js
@@ -4,6 +4,6 @@ class TaskListEngine extends Kensington {
   sortableList = this.createCustomTag('k-sortable-list', { onreorder: Function });
 }
 
-const t = new TaskListEngine({ validationLevel: 'error' });
+const t = new TaskListEngine({ debugMode: true, validationLevel: 'error' });
 
 export default t;